Use ChefDK instead of installing rubygems
The dependency graph became unwieldy to attempt to maintain, and many
guard plugins are net kept as up-to-date as the underlying libraries.

The kitchen-cloud plugins seemed a bit of an overload, since we are not
even testing against those cloud providers. If needed, this can be
re-instated at a later date where it is actually exercised.

Also leverage CircleCI parallelism instead of forcing it inside a single
build.